# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Receipt-to-Wallet API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List bank transactions
      parameters:
        - name: accountId
          in: query
          schema:
            type: ["string", "null"]
        - name: pending
          in: query
          schema:
            type: ["boolean", "null"]
        - name: from
          in: query
          schema:
            type: ["string", "null"]
          description: ISO date (YYYY-MM-DD)
        - name: to
          in: query
          schema:
            type: ["string", "null"]
          description: ISO date (YYYY-MM-DD)
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
...
  /transactions/{transactionId}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a single bank transaction
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── RECEIPTS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── ACCOUNT ──────────────────────────────────────────────────────────────
...
    Transaction:
      type: object
      required: [id, accountId, amount, currency, date, pending, createdAt]
      properties:
        id:
          type: string
        accountId:
          type: string
        amount:
          type: number
        currency:
          type: string
        merchantName:
          type: ["string", "null"]
        merchantNameRaw:
          type: ["string", "null"]
        categoryPrimary:
          type: ["string", "null"]
        categoryDetail:
          type: ["string", "null"]
        date:
          type: string
        pending:
          type: boolean
        matchId:
          type: ["integer", "null"]
        createdAt:
          type: string
...